backend/requirements.txt
fastapi
uvicorn
python-dotenv
web3
aiofiles
httpx
uuid
docker